Create core PASCal module
